List 디자인패턴정리 완료
